Update, Delete and Display snacks functionality

+++ Project/DataModels/SnacksModel.cs
@@ -5,7 +5,7 @@ public class SnacksModel
     public string Name { get; set; }
     public double Price { get; set; }
     public Int64 Id { get; set; }
-    public SnacksModel(Int64 id ,string name, float price)
+    public SnacksModel(Int64 id ,string name, double price)
     {
         Id = id;
         Name = name;
